+++ Packages/SublimeAStyleFormatter/AStyleFormat.py
@@ -60,7 +60,34 @@ def load_settings():
     return sublime.load_settings(PLUGIN_NAME + '.sublime-settings')
 
 
-def get_setting_for_view(view, key, default=None):
+_VARPROG_RE = re.compile(r'\$(\w+|\{[^}]*\})')
+
+
+def custom_expandvars(path, custom_envs):
+    if '$' not in path:
+        return path
+    envs = custom_envs.copy()
+    envs.update(os.environ)
+    i = 0
+    while True:
+        m = _VARPROG_RE.search(path, i)
+        if not m:
+            break
+        i, j = m.span(0)
+        name = m.group(1)
+        if name.startswith('{') and name.endswith('}'):
+            name = name[1:-1]
+        if name in envs:
+            tail = path[j:]
+            path = path[:i] + envs[name]
+            i = len(path)
+            path += tail
+        else:
+            i = j
+    return path
+
+
+def get_settings_for_view(view, key, default=None):
     try:
         settings = view.settings()
         sub_key = 'AStyleFormatter'
@@ -74,8 +101,9 @@ def get_setting_for_view(view, key, default=None):
     return settings.get(key, default)
 
 
-def get_setting_for_active_view(key, default=None):
-    return get_setting_for_view(sublime.active_window().active_view(), key, default)
+def get_settings_for_active_view(key, default=None):
+    return get_settings_for_view(
+        sublime.active_window().active_view(), key, default)
 
 
 def get_syntax_for_view(view):
@@ -87,7 +115,8 @@ def get_syntax_for_view(view):
 
 
 def is_supported_syntax(view, syntax):
-    mapping = get_setting_for_view(view, 'user_defined_syntax_mode_mapping', {})
+    mapping = get_settings_for_view(
+        view, 'user_defined_syntax_mode_mapping', {})
     return syntax in get_syntax_mode_mapping(mapping)
 
 
@@ -97,111 +126,168 @@ def is_enabled_in_view(view):
 
 
 class AstyleformatCommand(sublime_plugin.TextCommand):
-    def _get_setting(self, key, default=None):
-        return get_setting_for_view(self.view, key, default=default)
+    def _get_settings(self, key, default=None):
+        return get_settings_for_view(self.view, key, default=default)
 
-    def _get_syntax_setting(self, syntax, formatting_mode):
+    def _get_syntax_settings(self, syntax, formatting_mode):
         key = 'options_%s' % formatting_mode
-        setting = get_setting_for_view(self.view, key, default={})
+        settings = get_settings_for_view(self.view, key, default={})
         if syntax and syntax != formatting_mode:
             key = 'options_%s' % syntax
-            setting_override = get_setting_for_view(self.view, key, default={})
-            setting.update(setting_override)
-        return setting
+            settings_override = get_settings_for_view(
+                self.view, key, default={})
+            settings.update(settings_override)
+        return settings
 
     def _get_default_options(self):
         options_default = OPTIONS_DEFAULT.copy()
-        options_default_override = self._get_setting('options_default', default={})
+        options_default_override = self._get_settings(
+            'options_default', default={})
         options_default.update(options_default_override)
         return options_default
 
-    @staticmethod
-    def _read_astylerc(path):
+    _SKIP_COMMENT_RE = re.compile(r'\s*\#')
+
+    def _build_custom_vars(self):
+        view = self.view
+        custom_vars = {
+            'packages': sublime.packages_path(),
+        }
+        full_path = view.file_name()
+        if full_path:
+            file_name = os.path.basename(full_path)
+            file_base_name, file_extension = os.path.splitext(file_name)
+            custom_vars.update({
+                'file_path': os.path.dirname(full_path),
+                'file': full_path,
+                'file_name': file_name,
+                'file_extension': file_extension,
+                'file_base_name': file_base_name,
+            })
+        if sublime.version() > '3000':
+            window = view.window()
+            project_file_name = window.project_file_name()
+            if project_file_name:
+                project_name = os.path.basename(project_file_name)
+                project_base_name, project_extension = os.path.splitext(
+                    project_name)
+                custom_vars.update({
+                    'project': project_file_name,
+                    'project_path': os.path.dirname(project_file_name),
+                    'project_name': project_name,
+                    'project_extension': project_extension,
+                    'project_base_name': project_base_name,
+                })
+        return custom_vars
+
+    def _read_astylerc(self, path):
         # Expand environment variables first
-        fullpath = os.path.expandvars(path)
-        if not os.path.exists(fullpath) or not os.path.isfile(fullpath):
+        fullpath = custom_expandvars(path, self._build_custom_vars())
+        if not os.path.isfile(fullpath):
             return ''
         try:
-            skip_comment = re.compile(r'\s*\#')
             lines = []
             with open(fullpath, 'r') as f:
                 for line in f:
-                    if not skip_comment.match(line):
+                    if not self._SKIP_COMMENT_RE.match(line):
                         lines.append(line.strip())
             return ' '.join(lines)
-        except:
+        except Exception:
             return ''
-        return ''
 
     @staticmethod
     def _join_options(options_list):
-        options_string = ''
-        first = True
-        for i, o in enumerate(options_list):
-            if len(o) == 0:
-                continue
-            if first:
-                options_string += o
-                first = False
-            else:
-                options_string += ' ' + o
-        return options_string
+        return ' '.join(o for o in options_list if o)
 
-    def get_options(self, syntax, formatting_mode):
-        syntax_setting = self._get_syntax_setting(syntax, formatting_mode)
+    def _get_options(self, syntax, formatting_mode):
+        syntax_settings = self._get_syntax_settings(syntax, formatting_mode)
         # --mode=xxx placed first
-        options_list = [Options.parse_mode_setting(formatting_mode)]
+        options_list = [Options.build_astyle_mode_option(formatting_mode)]
 
-        if 'additional_options_file' in syntax_setting:
-            astylerc_string = self._read_astylerc(syntax_setting['additional_options_file'])
+        if 'additional_options_file' in syntax_settings:
+            astylerc_options = self._read_astylerc(
+                syntax_settings['additional_options_file'])
         else:
-            astylerc_string = ''
+            astylerc_options = ''
 
-        if 'additional_options' in syntax_setting:
-            additional_options = ' '.join(syntax_setting['additional_options'])
+        if 'additional_options' in syntax_settings:
+            additional_options = ' '.join(
+                syntax_settings['additional_options'])
         else:
             additional_options = ''
 
         options_list.append(additional_options)
-        options_list.append(astylerc_string)
+        options_list.append(astylerc_options)
 
-        # Check if user will use only additional options
-        # Skip processing other options when 'use_only_additional_options' is true
-        if syntax_setting.get('use_only_additional_options', False):
+        # Check if user will use only additional options, skip processing other
+        # options when 'use_only_additional_options' is true
+        if syntax_settings.get('use_only_additional_options', False):
             return self._join_options(options_list)
 
         # Get default options
-        default_setting = self._get_default_options()
-        # Merge syntax_setting with default_setting
-        default_setting.update(syntax_setting)
-        options = ' '.join(Options.parse_setting(default_setting))
+        default_settings = self._get_default_options()
+        # Merge syntax_settings with default_settings
+        default_settings.update(syntax_settings)
+        options = Options.build_astyle_options(
+            default_settings,
+            self._build_indent_options(),
+            convert_tabs=self._should_convert_tabs()
+        )
+        options = ' '.join(options)
         options_list.insert(1, options)
         return self._join_options(options_list)
 
+    def _build_indent_options(self):
+        view_settings = self.view.settings()
+        return {
+            'indent': 'spaces'
+                      if view_settings.get('translate_tabs_to_spaces')
+                      else 'tab',
+            'spaces': view_settings.get('tab_size'),
+        }
+
+    def _should_convert_tabs(self):
+        view_settings = self.view.settings()
+        return view_settings.get('translate_tabs_to_spaces')
+
     def _get_formatting_mode(self, syntax):
-        mapping = get_setting_for_view(self.view,
-                                       'user_defined_syntax_mode_mapping',
-                                       {})
+        mapping = get_settings_for_view(
+            self.view, 'user_defined_syntax_mode_mapping', {})
         return get_syntax_mode_mapping(mapping).get(syntax, '')
 
     def run(self, edit, selection_only=False):
+        # Close output panel previouslly created each run
+        under_unittest = self.view.settings().get('_UNDER_UNITTEST')
+        error_panel = ErrorMessagePanel("astyle_error_message",
+                                        under_unittest=under_unittest)
+        error_panel.close()
+
         try:
             # Loading options
             syntax = get_syntax_for_view(self.view)
             formatting_mode = self._get_formatting_mode(syntax)
-            options = self.get_options(syntax, formatting_mode)
-        except Options.RangeError as e:
-            sublime.error_message(str(e))
+            options = self._get_options(syntax, formatting_mode)
+        except Options.ImproperlyConfigured as e:
+            extra_message = e.extra_message
+            error_panel = ErrorMessagePanel("astyle_error_message")
+            error_panel.write(
+                "%s: An error occured while processing options: %s\n\n" % (
+                    PLUGIN_NAME, e))
+            if extra_message:
+                error_panel.write("* %s\n" % extra_message)
+            error_panel.show()
             return
         # Options ok, format now
         if selection_only:
             self.run_selection_only(edit, options)
         else:
             self.run_whole_file(edit, options)
-        if self._get_setting('debug', False):
+        if self._get_settings('debug', False):
             log_debug('AStyle version: {0}', pyastyle.version())
-            log_debug('Options: ' + options)
+            log_debug('AStyle options: ' + options)
         sublime.status_message('AStyle (v%s) Formatted' % pyastyle.version())
+
+    _STRIP_LEADING_SPACES_RE = re.compile(r'[ \t]*\n([^\r\n])')
 
     def run_selection_only(self, edit, options):
         def get_line_indentation_pos(view, point):
@@ -222,7 +308,8 @@ class AstyleformatCommand(sublime_plugin.TextCommand):
                 ch = view.substr(i)
                 scope = view.scope_name(i)
                 # Skip preprocessors, strings, characaters and comments
-                if 'string.quoted' in scope or 'comment' in scope or 'preprocessor' in scope:
+                if ('string.quoted' in scope or
+                        'comment' in scope or 'preprocessor' in scope):
                     extent = view.extract_scope(i)
                     i = extent.a - 1
                     continue
@@ -254,9 +341,11 @@ class AstyleformatCommand(sublime_plugin.TextCommand):
                 for _ in range(indent_count):
                     index = formatted_code.find('{') + 1
                     formatted_code = formatted_code[index:]
-                formatted_code = re.sub(r'[ \t]*\n([^\r\n])', r'\1', formatted_code, 1)
+                formatted_code = self._STRIP_LEADING_SPACES_RE.sub(
+                    r'\1', formatted_code, 1)
             else:
-                # HACK: While no identation, a '{' will generate a blank line, so strip it.
+                # HACK: While no identation, a '{' will generate a blank line,
+                # so strip it.
                 search = '\n{'
                 if search not in text:
                     formatted_code = formatted_code.replace(search, '{', 1)
@@ -264,9 +353,11 @@ class AstyleformatCommand(sublime_plugin.TextCommand):
             view.replace(edit, region, formatted_code)
             # Region for replaced text
             if sel.a <= sel.b:
-                regions.append(sublime.Region(region.a, region.a + len(formatted_code)))
+                regions.append(
+                    sublime.Region(region.a, region.a + len(formatted_code)))
             else:
-                regions.append(sublime.Region(region.a + len(formatted_code), region.a))
+                regions.append(
+                    sublime.Region(region.a + len(formatted_code), region.a))
         view.sel().clear()
         # Add regions of formatted text
         [view.sel().add(region) for region in regions]
@@ -280,7 +371,8 @@ class AstyleformatCommand(sublime_plugin.TextCommand):
         # Replace to view
         _, err = merge_code(view, edit, code, formatted_code)
         if err:
-            sublime.error_message('%s: Merge failure: "%s"' % (PLUGIN_NAME, err))
+            error_panel = ErrorMessagePanel("astyle_error_message")
+            error_panel.write('%s: Merge failure: "%s"\n' % (PLUGIN_NAME, err))
 
     def is_enabled(self):
         return is_enabled_in_view(self.view)
@@ -288,10 +380,54 @@ class AstyleformatCommand(sublime_plugin.TextCommand):
 
 class PluginEventListener(sublime_plugin.EventListener):
     def on_pre_save(self, view):
-        if is_enabled_in_view(view) and get_setting_for_active_view('autoformat_on_save', default=False):
+        if is_enabled_in_view(view) and get_settings_for_active_view(
+                'autoformat_on_save', default=False):
             view.run_command('astyleformat')
 
     def on_query_context(self, view, key, operator, operand, match_all):
         if key == 'astyleformat_is_enabled':
             return is_enabled_in_view(view)
         return None
+
+
+class AstylePanelInsertCommand(sublime_plugin.TextCommand):
+
+    def run(self, edit, text):
+        self.view.set_read_only(False)
+        self.view.insert(edit, self.view.size(), text)
+        self.view.set_read_only(True)
+        self.view.show(self.view.size())
+
+
+class ErrorMessagePanel(object):
+    def __init__(self, name, under_unittest=False, word_wrap=False,
+                 line_numbers=False, gutter=False, scroll_past_end=False,
+                 syntax='Packages/Text/Plain text.tmLanguage'):
+        # If we are under testing, do not manipulate output panel
+        self.name = name
+        self.window = None
+        self.output_view = None
+        if not under_unittest:
+            self.window = sublime.active_window()
+            self.output_view = self.window.get_output_panel(name)
+
+            settings = self.output_view.settings()
+            settings.set("word_wrap", word_wrap)
+            settings.set("line_numbers", line_numbers)
+            settings.set("gutter", gutter)
+            settings.set("scroll_past_end", scroll_past_end)
+            settings.set("syntax", syntax)
+
+    def write(self, s):
+        if self.output_view:
+            self.output_view.run_command('astyle_panel_insert', {'text': s})
+
+    def show(self):
+        if self.output_view:
+            self.window.run_command(
+                "show_panel", {"panel": "output." + self.name})
+
+    def close(self):
+        if self.output_view:
+            self.window.run_command(
+                "hide_panel", {"panel": "output." + self.name})